+++ packages/core/core/lib/middlewares/error-handler/index.js
@@ -1,15 +1,15 @@
 /*
  * Author: Kevin Ries (kevin@fachw3rk.de)
  * -----
- * Copyright 2020 Fachwerk
+ * Copyright 2021 Fachwerk
  */
 
 const { WeaveError } = require('../../errors')
 
 module.exports = (runtime) => {
   const wrapErrorHandlerMiddleware = function (handler) {
-    return function errorHandlerMiddleware (context) {
-      return handler(context)
+    return function errorHandlerMiddleware (context, serviceInjections) {
+      return handler(context, serviceInjections)
         .catch(error => {
           if (!(error instanceof Error)) {
             error = new WeaveError(error, 500)