+++ packages/core/core/lib/middlewares/tracing/index.js
@@ -1,7 +1,7 @@
 /*
  * Author: Kevin Ries (kevin@fachw3rk.de)
  * -----
- * Copyright 2020 Fachwerk
+ * Copyright 2021 Fachwerk
  */
 
 const { buildTags } = require('./buildTags')
@@ -11,7 +11,7 @@ const wrapTracingLocalActionMiddleware = function (handler) {
   const options = broker.options.tracing || {}
 
   if (options.enabled) {
-    return function metricsLocalMiddleware (context) {
+    return function metricsLocalMiddleware (context, serviceInjections) {
       const tags = buildTags(context)
 
       const spanName = `action '${context.action.name}'`
@@ -28,7 +28,7 @@ const wrapTracingLocalActionMiddleware = function (handler) {
 
       context.span = span
 
-      return handler(context)
+      return handler(context, serviceInjections)
         .then(result => {
           span.finish()
           return result