+++ packages/core/repl/lib/commands/call.js
@@ -4,9 +4,9 @@ module.exports = ({ vorpal, broker }) => {
   vorpal
     .command('call <actionName> [jsonParams]', 'Call an action.')
     .alias('c')
-    .option('--l [filename]', 'Load params from file')
+    .option('-f, --file [filename]', 'Load params from file')
     .option('--stream [filename]', 'Send a file as stream')
-    .option('--s [filename]', 'Save response to file')
+    .option('-s, --save [filename]', 'Save response to file')
     .autocomplete({
       data () {
         return [...new Set(broker.runtime.registry.actionCollection.list({}).map(item => item.name))]