+++ packages/cache/redis/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.10.0](https://github.com/weave-microservices/weave/compare/@weave-js/redis-cache@0.9.4...@weave-js/redis-cache@0.10.0) (2021-10-18)
+
+**Note:** Version bump only for package @weave-js/redis-cache
+
+
+
+
+
 ## [0.9.4](https://github.com/weave-microservices/weave/compare/@weave-js/redis-cache@0.9.3...@weave-js/redis-cache@0.9.4) (2021-09-09)
 
 **Note:** Version bump only for package @weave-js/redis-cache