+++ packages/core/repl/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.10.0](https://github.com/weave-microservices/weave/compare/@weave-js/repl@0.9.5...@weave-js/repl@0.10.0) (2021-10-18)
+
+
+### Bug Fixes
+
+* Event balancing ([6fa2b58](https://github.com/weave-microservices/weave/commit/6fa2b58e50a8fd471c63e5a5678831e7986b92d0))
+
+
+
+
+
 ## [0.9.5](https://github.com/weave-microservices/weave/compare/@weave-js/repl@0.9.4...@weave-js/repl@0.9.5) (2021-09-09)
 
 **Note:** Version bump only for package @weave-js/repl